python/fastapi-logger/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Duc Nguyen <vanducld@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
dev = 'server:main'

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.24.0.post1"
fastapi = "^0.105.0"
python-dotenv = "^1.0.0"
watchtower = "^3.0.1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
